Report an error if we can't get the workflow artifacts list
It seems we had an incident where the metadata file wasn't retrieved, so
an exception later occurred as we didn't know the buildnumber. Instead,
raise an exception if fetching the artifacts list fails.

That kind of failure could also occur if we succeed in fetching the
artifact list, but the metadata file isn't in it. Mark the job as
'errored' in that case, since something's gone horribly wrong.