Make a separate source package build job possible
At the moment, this just runs 'cygport srcpackage' to make the source
package. Later we will provide it as an artifact to the install package
builds, rather than checking out the packaging repository.

This will be an improvement because:

* This detects when the source package isn't actually buildable due to
files which are present in the source repository but not listed in
SRC_URI etc.

* If the build system for the package looks to see if it's being built
in a git repository, this prevents it finding the packaging respository
and perhaps using an inappropriate commit id or tag (or getting confused
by unexpected tag name pattern or a lack of tags).