Easier debugging on a test instance
Make it possible to invoke the GitHub repository_dispatch API on a test
instance of this repository for testing and debugging.

Set the GITHUB_DEBUG_OWNER env var to the owner of the test instance,
and provide an access token in the file 'github.token'.

Then invoking post-recieve with a suitable (fake) context will be
directed by the request_build module at the test instance, e.g:

  export GL_REPO="git/cygwin-packages/$PACKAGE"
  export CYGNAME="$MAINTAINER"
  echo $COMMIT $COMMIT $REFERENCE | ./post-receive