Stop using setup flag '-g' ('--upgrade-also')
We added this in 2c4c60bd, to work around a bug in setup, which was
subsequently fixed.

Using this flag has the unfortunate side-effect that if a package
containing a DLL we have loaded is updated on a mirror between the
install we did to run scallywag, and the install of the build
requirements, that install will try to update that package, and fail, as
the file is in use.