Pass source package to arch build(s)
(On AppVeyor, we keep on doing everything in a single job. Passing an
artifact between jobs seems like something we'd have to cook up
ourselves, and that seems like a lot of work for little benefit)

This contains a change to the fetcher to present the upload with
separated source in the src/, which the uploader needs to be prepared to
accept.